requirements.txt
numpy>=1.16.4
scipy>=1.1.0
torch>=1.8.0
asteroid-filterbanks>=0.4.0
requests
filelock
SoundFile>=0.10.2
huggingface_hub>=0.0.2
PyYAML>=5.0
pandas>=0.23.4
pytorch-lightning>=1.0.1
torchaudio>=0.8.0
pb_bss_eval>=0.0.2
torch_stoi>=0.0.1
torch_optimizer>=0.0.1a12
julius